written basic test for morse translator no error checks yet
