+++ index.html
@@ -4,13 +4,13 @@
         <meta charset="UTF-8" />
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <link rel="stylesheet" href="./styles.css" />
-        <script type="module" src="./script.js" defer></script>
+        <link rel="stylesheet" href="./styles/styles.css" />
+        <script type="module" src="./scripts/script.js" defer></script>
         <title>Document</title>
     </head>
     <body>
         <form>
-            <label for="toTranslateInput">English</label>
+            <label for="toTranslateInput">Auto Detect</label>
             <textarea
                 id="toTranslateInput"
                 cols="30"
@@ -20,10 +20,9 @@
             ></textarea>
             <input type="submit" id="translateButton" value="translate" />
         </form>
-        <!-- <button id="toggleTranslationBtn">Toggle Translation</button> -->
 
         <div class="translated-container">
-            <h2 class="tanslated-container__heading">Morse Code</h2>
+            <h2 class="tanslated-container__heading">Auto Detect</h2>
             <p
                 id="outputArea"
                 class="translated-container__translated-text"
