--- script.test.js
@@ -1,5 +1,0 @@
-describe("test that tests working", () => {
-    it("tests it's working", () => {
-        expect(true).toBe(true);
-    });
-});
