+++ scripts/package.json
@@ -1,5 +1,8 @@
 {
     "devDependencies": {
+        "@babel/core": "^7.21.4",
+        "@babel/preset-env": "^7.21.4",
+        "babel-jest": "^29.5.0",
         "jest": "^29.5.0"
     },
     "scripts": {
